requirements.txt
gspread
oauth2client
python_binance
flask
Flask-SQLAlchemy
flask-login
PyPDF2
borb